Minor: Move `Partitioning` and `Distribution` to `physical_expr`
